fix(decorators): wait for lime properties in componentWillLoad
When a web component is created by Angular it is inserted into the DOM before any props are set.
This will cause the state decorators to fail since the platform prop will be undefined. It will
always work the first time since the components are lazy loaded and Angular will have time to set
the props before the lifecycle hooks run, but if any more components are created after that it will
fail. This change makes sure that the props are set before the lifecycle hooks run.

fix Lundalogik/crm-feature#1004